requirements-dev.txt
-r requirements.txt

mypy==1.9.0
pytest==8.2.0
testing.postgresql==1.3.0
